Implemented storing of multiple images against user account, at loss of delete functionality. Facilitated offline development through blobstore-type 'get_serving_url'. Added signin page

+++ delete.py
@@ -4,7 +4,10 @@ sys.path.append('lib')
 
 import crud
 import models
+import time
 import webapp2
+
+from google.appengine.api import users
 
 # [END imports]
 
@@ -14,14 +17,22 @@ class DeletePage(webapp2.RequestHandler):
     # Handle delete post action
     def get(self):
 
+        imageid = self.request.get('imageid')
+
+        # Retrieve user
+        user = users.get_current_user()
+
         # Retrieve image model
-        user_image = models.get_user_image()
+        user_image = models.Image.get_by_id(imageid)
 
-        # Delete image from cloud cloud storage
-        crud.delete_file('/' + user_image.bucket_name + '/' + user_image.name)
 
-        # Delete metadata from datastore
-        models.delete_user_image()
+        if user_image.user_id == user.user_id():
+
+            # Delete image from cloud cloud storage
+            crud.delete_file('/' + user_image.bucket_name + '/' + user_image.name)
+
+            # Delete metadata from datastore
+            models.Image.delete_user_image(imageid)
 
         # Return to main page
         self.redirect('/')
